fix: revise 0322 - Coin Change - Medium

+++ src/problems/0301-0400/0322_coin-change.js
@@ -44,7 +44,7 @@ const coinChange = (coins, amount) => {
                 return 0;
         }
 
-        const v = new Uint16Array(amount + 1);
+        const v = new Uint8Array(amount + 1);
         v[amount] = 1;
         let q = [amount];
         let res = 1;
